* origin/sherlock:
  Fix to the sporadic failures in e2e tests.
  e2e add async compact pattern
  use snapshots when creating checkpoints
  Correct the durability option to FDB_DRB_NONE for normal_commit_autocompact_test
  Increase file ref_count to 4 bytes to prevent overflow crashes
  MB-13955: avoid looping on random picks for bcache_evicts

+++ src/filemgr.h
@@ -97,7 +97,7 @@ struct fnamedic_item;
 struct kvs_header;
 struct filemgr {
     char *filename; // Current file name.
-    uint8_t ref_count;
+    uint32_t ref_count;
     uint8_t fflags;
     uint16_t filename_len;
     uint32_t blocksize;
